refactor(cartItemProvider): 내부의 훅 분리
